# openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Detail"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Call": {
        "additionalProperties": false,
        "properties": {
          "answered_at": {
            "description": "Not recorded for app calls today: null.",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "cost": {
            "description": "The recorded charge, decimal string with two decimals; null unless status is completed.",
            "type": [
              "string",
              "null"
            ]
          },
          "counterparty_name": {
            "description": "Name of the contact whose stored phone number exactly matches the other party (to_number outbound, from_number inbound); a contact saved in another format, such as with spaces, does not match. Requires contacts:read, else null.",
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "enum": [
              "USD"
            ],
            "type": "string"
          },
          "direction": {
            "enum": [
              "outbound",
              "inbound"
            ],
            "type": "string"
          },
          "duration_seconds": {
            "description": "Billable seconds once the call has ended.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "ended_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "from_number": {
            "description": "E.164 caller ID used.",
            "type": "string"
          },
          "id": {
            "pattern": "^call_[0-9a-f]{32}$",
            "type": "string"
          },
          "is_recorded": {
            "type": "boolean"
          },
          "livemode": {
            "type": "boolean"
          },
          "object": {
            "description": "Always \"call\".",
            "enum": [
              "call"
            ],
            "type": "string"
          },
          "origin": {
            "description": "Which surface placed the call.",
            "enum": [
              "app",
              "api",
              "mcp"
            ],
            "type": "string"
          },
          "placement": {
            "description": "How the call was set up: from the app or browser (webrtc), by ringing your phone first (callback), or received on your number (inbound).",
            "enum": [
              "webrtc",
              "callback",
              "inbound"
            ],
            "type": "string"
          },
          "rate_per_minute": {
            "description": "The per-minute rate fixed at call setup and charged for this call, decimal string with four decimals.",
            "pattern": "^[0-9]+\\.[0-9]{4}$",
            "type": "string"
          },
          "session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CallSession"
              },
              {
                "type": "null"
              }
            ],
            "description": "Callback calls only; null for app and inbound calls."
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "description": "initiating, connecting, active, completed, failed, no-answer, busy, canceled; callback calls add ringing_you and ringing_them. New values may appear.",
            "type": "string"
          },
          "to_number": {
            "description": "E.164.",
            "type": "string"
          },
          "transcript_status": {
            "description": "null when no transcript was requested.",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              null
            ],
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "object",
          "livemode",
          "direction",
          "placement",
          "origin",
          "status",
          "to_number",
          "from_number",
          "counterparty_name",
          "started_at",
          "answered_at",
          "ended_at",
          "duration_seconds",
          "rate_per_minute",
          "currency",
          "cost",
          "is_recorded",
          "transcript_status",
          "session"
        ],
        "type": "object"
      },
      "CallDetail": {
        "additionalProperties": false,
        "properties": {
          "answered_at": {
            "description": "Not recorded for app calls today: null.",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "cost": {
            "description": "The recorded charge, decimal string with two decimals; null unless status is completed.",
            "type": [
              "string",
              "null"
            ]
          },
          "cost_so_far": {
            "description": "Charge accrued so far, decimal string, only while status is active; null otherwise (including initiating and connecting). Counted from started_at, so it includes ring time until answered_at is recorded.",
            "type": [
              "string",
              "null"
            ]
          },
          "counterparty_name": {
            "description": "Name of the contact whose stored phone number exactly matches the other party (to_number outbound, from_number inbound); a contact saved in another format, such as with spaces, does not match. Requires contacts:read, else null.",
            "type": [
              "string",
              "null"
            ]
          },
          "currency": {
            "enum": [
              "USD"
            ],
            "type": "string"
          },
          "direction": {
            "enum": [
              "outbound",
              "inbound"
            ],
            "type": "string"
          },
          "duration_seconds": {
            "description": "Billable seconds once the call has ended.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "ended_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "from_number": {
            "description": "E.164 caller ID used.",
            "type": "string"
          },
          "id": {
            "pattern": "^call_[0-9a-f]{32}$",
            "type": "string"
          },
          "is_recorded": {
            "type": "boolean"
          },
          "livemode": {
            "type": "boolean"
          },
          "max_duration_seconds": {
            "description": "Time limit the balance allowed when the call was set up, while status is active. Derived from the balance, so it needs credits:read; null without it. Not recorded for app-placed personal calls today, so null for them.",
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "object": {
            "description": "Always \"call\".",
            "enum": [
              "call"
            ],
            "type": "string"
          },
          "origin": {
            "description": "Which surface placed the call.",
            "enum": [
              "app",
              "api",
              "mcp"
            ],
            "type": "string"
          },
          "placement": {
            "description": "How the call was set up: from the app or browser (webrtc), by ringing your phone first (callback), or received on your number (inbound).",
            "enum": [
              "webrtc",
              "callback",
              "inbound"
            ],
            "type": "string"
          },
          "rate_per_minute": {
            "description": "The per-minute rate fixed at call setup and charged for this call, decimal string with four decimals.",
            "pattern": "^[0-9]+\\.[0-9]{4}$",
            "type": "string"
          },
          "session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CallSession"
              },
              {
                "type": "null"
              }
            ],
            "description": "Callback calls only; null for app and inbound calls."
          },
          "started_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "description": "initiating, connecting, active, completed, failed, no-answer, busy, canceled; callback calls add ringing_you and ringing_them. New values may appear.",
            "type": "string"
          },
          "to_number": {
            "description": "E.164.",
            "type": "string"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Transcript"
              },
              {
                "type": "null"
              }
            ],
            "description": "Always null today. Reserved for the call summary once reading summaries through connected apps can be enabled (Settings → Connected apps); until then expand[]=transcript answers 403 summaries_not_enabled."
          },
          "transcript_status": {
            "description": "null when no transcript was requested.",
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed",
              null
            ],
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "cost_so_far",
          "max_duration_seconds",
          "transcript",
          "id",
          "object",
          "livemode",
          "direction",
          "placement",
          "origin",
          "status",
          "to_number",
          "from_number",
          "counterparty_name",
          "started_at",
          "answered_at",
          "ended_at",
          "duration_seconds",
          "rate_per_minute",
          "currency",
          "cost",
          "is_recorded",
          "transcript_status",
          "session"
        ],
        "type": "object"
      },
      "CallSession": {
        "additionalProperties": false,
        "properties": {
          "hangup_cause": {
            "type": [
              "string",
              "null"
            ]
          },
          "leg_a_status": {
            "type": "string"
          },
          "leg_b_status": {
            "type": "string"
          },
          "ring_number": {
            "description": "The verified number that was rung.",
            "type": "string"
          }
        },
        "required": [
          "ring_number",
          "leg_a_status",
          "leg_b_status",
          "hangup_cause"
        ],
        "type": "object"
      },
...
      "Detail": {
        "additionalProperties": false,
        "properties": {
          "billing_url": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "doc_url": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "param": {
            "type": [
              "string",
              "null"
            ]
          },
          "request_id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "code",
          "message",
          "param",
          "doc_url",
          "request_id"
        ],
        "type": "object"
      },
      "ListCall": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Call"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "has_more": {
            "description": "Whether more objects exist beyond this page in the direction of travel.",
            "type": "boolean"
          },
          "object": {
            "description": "Always \"list\".",
            "enum": [
              "list"
            ],
            "type": "string"
          }
        },
        "required": [
          "object",
          "data",
          "has_more"
        ],
        "type": "object"
      },
...
      "ReferenceNumber": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "label",
          "value"
        ],
        "type": "object"
      },
...
      "Transcript": {
        "additionalProperties": false,
        "properties": {
          "action_items": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "call_id": {
            "pattern": "^call_[0-9a-f]{32}$",
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "duration_seconds": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "pattern": "^tr_[0-9a-f]{32}$",
            "type": "string"
          },
          "language": {
            "description": "BCP 47 tag when detected.",
            "type": [
              "string",
              "null"
            ]
          },
          "livemode": {
            "type": "boolean"
          },
          "object": {
            "description": "Always \"transcript\".",
            "enum": [
              "transcript"
            ],
            "type": "string"
          },
          "reference_numbers": {
            "items": {
              "$ref": "#/components/schemas/ReferenceNumber"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "type": "string"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "object",
          "livemode",
          "call_id",
          "status",
          "summary",
          "action_items",
          "reference_numbers",
          "language",
          "duration_seconds",
          "completed_at"
        ],
        "type": "object"
      },
      "User": {
        "additionalProperties": false,
        "properties": {
          "country_code": {
            "description": "ISO 3166-1 alpha-2 country, when known. Not stored yet: always null.",
            "type": [
              "string",
              "null"
            ]
          },
          "display_name": {
            "description": "Display name.",
            "type": "string"
          },
          "email": {
            "description": "Sign-in email.",
            "format": "email",
            "type": "string"
          },
          "id": {
            "description": "Account id (usr_ prefix).",
            "pattern": "^usr_[0-9a-f]{32}$",
            "type": "string"
          },
          "livemode": {
            "description": "true for OAuth access tokens and live API keys.",
            "type": "boolean"
          },
          "object": {
            "description": "Always \"user\".",
            "enum": [
              "user"
            ],
            "type": "string"
          },
          "time_zone": {
            "description": "IANA time zone, when known. Not stored yet: always null.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "object",
          "id",
          "email",
          "display_name",
          "country_code",
          "time_zone",
          "livemode"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "Personal API key (vc_live_ or vc_test_) sent as a bearer token; /v1 only. Self-serve key creation is not available yet.",
        "scheme": "bearer",
        "type": "http"
      },
      "oauth2": {
        "description": "OAuth 2.1 authorization code with PKCE (S256). Public clients register by Client ID Metadata Document or Dynamic Client Registration.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://api.voixcall.com/oauth/authorize",
            "refreshUrl": "https://api.voixcall.com/oauth/token",
            "scopes": {
              "calls:place": "Place calls that ring your phone and are charged to your balance",
              "calls:read": "See your calls: numbers, times and cost",
              "contacts:read": "Search your contacts and show contact names on your calls",
              "credits:read": "See your credit balance and history",
              "messages:history": "Read all text messages on numbers you opt in",
              "messages:read": "Read text messages received in the last 10 minutes on numbers you opt in, including login codes",
              "numbers:read": "See your VoixCall numbers",
              "rates:read": "Look up what a call costs",
              "transcripts:read": "Read summaries and transcripts of your recorded calls",
              "webhooks:manage": "Manage webhook endpoints (API keys only)"
            },
            "tokenUrl": "https://api.voixcall.com/oauth/token",
            "x-usePkce": "SHA-256"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "contact": {
      "email": "support@voixcall.com",
      "name": "VoixCall support",
      "url": "https://voixcall.com/developers"
    },
    "description": "Rates, credits, calls, contacts, numbers and caller ID for the signed-in VoixCall account. Personal account only; organization data is not reachable.",
    "termsOfService": "https://voixcall.com/terms",
    "title": "VoixCall API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/calls": {
      "get": {
        "description": "Calls on the personal account, newest first. counterparty_name is filled from contacts only with contacts:read. Organization calls are never included. Test keys get an empty list. Requires calls:read.",
        "operationId": "list-calls",
        "parameters": [
          {
            "description": "Page size, 1 to 100. Default 10.",
            "explode": false,
            "in": "query",
            "name": "limit",
            "schema": {
              "description": "Page size, 1 to 100. Default 10.",
              "type": "string"
            }
          },
          {
            "description": "Return objects after this id (older).",
            "explode": false,
            "in": "query",
            "name": "starting_after",
            "schema": {
              "description": "Return objects after this id (older).",
              "type": "string"
            }
          },
          {
            "description": "Return objects before this id (newer).",
            "explode": false,
            "in": "query",
            "name": "ending_before",
            "schema": {
              "description": "Return objects before this id (newer).",
              "type": "string"
            }
          },
          {
            "description": "Only calls in this direction.",
            "explode": false,
            "in": "query",
            "name": "direction",
            "schema": {
              "description": "Only calls in this direction.",
              "enum": [
                "outbound",
                "inbound"
              ],
              "type": "string"
            }
          },
          {
            "description": "Only calls started at or after this RFC 3339 time.",
            "explode": false,
            "in": "query",
            "name": "started_at[gte]",
            "schema": {
              "description": "Only calls started at or after this RFC 3339 time.",
              "examples": [
                "2026-09-01T00:00:00Z"
              ],
              "type": "string"
            }
          },
          {
            "description": "Only calls started at or before this RFC 3339 time.",
            "explode": false,
            "in": "query",
            "name": "started_at[lte]",
            "schema": {
              "description": "Only calls started at or before this RFC 3339 time.",
              "type": "string"
            }
          },
          {
            "description": "Only calls to this number, E.164 exact match.",
            "explode": false,
            "in": "query",
            "name": "to",
            "schema": {
              "description": "Only calls to this number, E.164 exact match.",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCall"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [
              "calls:read"
            ]
          },
          {
            "apiKey": []
          }
        ],
        "summary": "List calls",
        "tags": [
          "calls"
        ]
      }
    },
    "/calls/{id}": {
      "get": {
        "description": "One call by id. While the call is active, cost_so_far is set (counted from started_at, so it includes ring time until answered_at is recorded), and max_duration_seconds with credits:read. transcript is always null: reading call summaries through connected apps is not enabled yet, so expand[]=transcript answers 403 summaries_not_enabled. A call that is not yours is not found. Requires calls:read.",
        "operationId": "get-call",
        "parameters": [
          {
            "description": "Call id (call_ prefix).",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "description": "Call id (call_ prefix).",
              "type": "string"
            }
          },
          {
            "description": "expand[]=transcript is reserved for the call summary. Reading summaries through connected apps is not enabled yet: with transcripts:read it answers 403 summaries_not_enabled; without it, 403 insufficient_scope.",
            "explode": true,
            "in": "query",
            "name": "expand[]",
            "schema": {
              "description": "expand[]=transcript is reserved for the call summary. Reading summaries through connected apps is not enabled yet: with transcripts:read it answers 403 summaries_not_enabled; without it, 403 insufficient_scope.",
              "items": {
                "enum": [
                  "transcript"
                ],
                "type": "string"
              },
              "type": [
                "array",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Not Found"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": [
              "calls:read"
            ]
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Get a call",
        "tags": [
          "calls"
        ]
      }
    },
...
    "/me": {
      "get": {
        "description": "Returns the VoixCall account the token or key acts for. Any scope.",
        "operationId": "get-me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Unauthorized"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "oauth2": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "Get the connected account",
        "tags": [
          "account"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.voixcall.com/v1"
    }
  ]
}